+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Upgrade dependencies
 + Refactor BankAccountTransactionsDTO for API post
++ Refactor repository package test in repositories
 
 ## 0.13.0
 + Add `BalanceNotEnoughException`